ai-customer-service-python/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic
python-multipart==0.0.6

# AI框架与模型
langchain==0.1.0
langchain-community==0.0.10
dashscope==1.14.0

# 向量数据库
chromadb==0.4.18
langchain-chroma

# 文档处理
pypdf==3.17.1
unstructured==0.11.0

# 配置与工具
pyyaml==6.0.1
redis==5.0.1
pandas==2.1.4

# Web界面
streamlit==1.28.0

# 其他
httpx==0.25.2
numpy==1.26.2
